Change filter for value default
With the standard jinja2 'default' filter a value of 0 would not be displayed.
Setting the default filter to 'default("something",False) would display the 0
but would also display 'None' if the field was null.

Custom filter default_if_none is very similar to builtin default but substitutes
the default value on None but not if the value evaluates to False.