Ensure that numeric fields get numeric values
Sqlite allows strings to be stored in any field regardless of the declared
type.

Storing non-numeric strings in numeric fields could cause problems.

Function _text_to_numeric(self,field_name,value,column_type) trys to coerce
the string to a number of the appropreate type.

Returns the numeric value or the original string if not successful.

The save() function actually calls _text_to_numeric as needed and also
tests any numeric values in the record to ensure they are the correct
type for the field definition