Add option to add new columns to a table
This amounts to a very simple migration tool. It only allows for
adding columns to a table after it is created. To use it modify the
table class definition as follows:

1. Add the method '_column_list(self)' with the @property decorator.
The method returns a list of dicts with the keys 'name' and
'definition' which will be used to add the column to the table if
it does not exist.

2. Add or modify the 'create_table' method of the class by adding
a second parameter as 'super().create_table(sql,self._column_list)'