Add record list export and defaults
Add the ability to export the current selection of records from
the list view. Default is to export the fields displayed in the view
but setting TableView.export_fields will define the fields to export.
Exported file is in .csv format

Also add 'default' element to TableView.list_fields property to set
the default value to display in the list view.