Add 'list' Property to generic view
Seting the 'list' property of a list field definition to False
will cause the field to be used in the search widget (assuming, that
is not set to False) but it will not be displayed in the list table.
Allows you to search in fields that you don't want to add to list.