Handle user/set_list_roles/ post better
For reasons I don't know the mutiple select form element
name submitted by the ajax request has "[]" appended to the
form element names in request.form instead of the actual name of the element.
I don't know if this is something jquery is doing or the browser or
something I'm doing but I test the values more carefully now.