Handle checkbox input differently
For some reason checkbox input elements that have the value
set to 1 are being checked by default. So I'm now setting the
value to some string and then I have to explicitly set the value
in the rec.<checkbox> data record because I use an integer value
for most check boxes.