No longer using Flask-Mail
Copied flask_mail.py into takeabeltof.mail.py and reworked it to work as
a normal module instead of a flask plugin.

The goal is to modify the code to allow for the app to log in to gmail
using Google oAuth credentials which flask_mail does not accomidate.

At this point gmail login is still not enabled. That is the next step.

It's not necessary to uninstall flask_mail or remove it from app.py to be
compatible with this change. It's simply not used anymore.