Add option to use Gmail API to send emails
The hope is that using this method I will be able to send mail to
gmail and gsuite accounts more reliably.

Google seems to be very fussy about who's mail they will forward to
their clients. In the little testing I've done so far, this seems to work
pretty reliably.