Fix error when updating User table
When updating the user record in users.models.update() there
is a special case where if 'active' is in the form, the form
was converted to an instance of rec. In the namedlist module
rec was iterable, but in my new dataclasses version RowData is not
and it throughs an error in SqliteTable.update().

Since the docs say that the form should be a dict like object, just
convert form to a dict before calling update()